pyproject.toml
[project]
name = "darams"
version = "0.1.0"
description = "Drift-aware Real-time Alpha Monitoring and Adaptation System based on DolphinDB WorldQuant 101 Alpha"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.25",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "dolphindb>=1.30",
    "pandas>=2.1",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "lightgbm>=4.2",
    "mlflow>=2.10",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "httpx>=0.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
]
notebook = [
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
